Inital Upload
These are the openscad models and their .stl renderings that have been
used so far for this project. These have been printed either in ABS or
PLA. The case files form a press-fit, non-watertight enclosure for our
board with cutouts for usb host/client, Ethernet and power. There is
enough clearance on the bottom of the case to allow our PCI Wifi card to
be attached and enough clearance ontop to allow mount of a PCB board to
jump our sensor connections and hold our LED's. The stand Holds the
webcam and mounts the PIR sensor just below it.